made a separate file for the Path class from the astar_search
